apps/requirements.txt
flask ~= 1.1
livereload

ffmpeg-python

google-cloud
google-cloud-storage
google-cloud-speech
google-api-python-client